prepare to replace existing implementation
